+++ sdk/resourcemanager/monitor/armmonitor/zz_generated_models.go
@@ -150,6 +150,13 @@ type ActionGroupResource struct {
 	Type *string `json:"type,omitempty" azure:"ro"`
 }
 
+// ActionGroupsClientBeginCreateNotificationsAtResourceGroupLevelOptions contains the optional parameters for the ActionGroupsClient.BeginCreateNotificationsAtResourceGroupLevel
+// method.
+type ActionGroupsClientBeginCreateNotificationsAtResourceGroupLevelOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
 // ActionGroupsClientBeginPostTestNotificationsOptions contains the optional parameters for the ActionGroupsClient.BeginPostTestNotifications
 // method.
 type ActionGroupsClientBeginPostTestNotificationsOptions struct {
@@ -174,6 +181,12 @@ type ActionGroupsClientEnableReceiverOptions struct {
 
 // ActionGroupsClientGetOptions contains the optional parameters for the ActionGroupsClient.Get method.
 type ActionGroupsClientGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// ActionGroupsClientGetTestNotificationsAtResourceGroupLevelOptions contains the optional parameters for the ActionGroupsClient.GetTestNotificationsAtResourceGroupLevel
+// method.
+type ActionGroupsClientGetTestNotificationsAtResourceGroupLevelOptions struct {
 	// placeholder for future optional parameters
 }
 
@@ -3795,18 +3808,6 @@ type TestNotificationDetailsResponse struct {
 
 	// The created time
 	CreatedTime *string `json:"CreatedTime,omitempty"`
-}
-
-// TestNotificationResponse - The response when test notification succeeded
-type TestNotificationResponse struct {
-	// REQUIRED; The correlation id
-	CorrelationID *string `json:"correlationId,omitempty"`
-
-	// REQUIRED; The created time
-	CreatedTime *string `json:"createdTime,omitempty"`
-
-	// REQUIRED; The notification id
-	NotificationID *string `json:"notificationId,omitempty"`
 }
 
 // ThresholdRuleCondition - A rule condition based on a metric crossing a threshold.