+++ sdk/resourcemanager/monitor/armmonitor/zz_generated_response_types.go
@@ -8,6 +8,11 @@
 
 package armmonitor
 
+// ActionGroupsClientCreateNotificationsAtResourceGroupLevelResponse contains the response from method ActionGroupsClient.CreateNotificationsAtResourceGroupLevel.
+type ActionGroupsClientCreateNotificationsAtResourceGroupLevelResponse struct {
+	// placeholder for future response values
+}
+
 // ActionGroupsClientCreateOrUpdateResponse contains the response from method ActionGroupsClient.CreateOrUpdate.
 type ActionGroupsClientCreateOrUpdateResponse struct {
 	ActionGroupResource
@@ -28,6 +33,11 @@ type ActionGroupsClientGetResponse struct {
 	ActionGroupResource
 }
 
+// ActionGroupsClientGetTestNotificationsAtResourceGroupLevelResponse contains the response from method ActionGroupsClient.GetTestNotificationsAtResourceGroupLevel.
+type ActionGroupsClientGetTestNotificationsAtResourceGroupLevelResponse struct {
+	TestNotificationDetailsResponse
+}
+
 // ActionGroupsClientGetTestNotificationsResponse contains the response from method ActionGroupsClient.GetTestNotifications.
 type ActionGroupsClientGetTestNotificationsResponse struct {
 	TestNotificationDetailsResponse
@@ -45,7 +55,7 @@ type ActionGroupsClientListBySubscriptionIDResponse struct {
 
 // ActionGroupsClientPostTestNotificationsResponse contains the response from method ActionGroupsClient.PostTestNotifications.
 type ActionGroupsClientPostTestNotificationsResponse struct {
-	TestNotificationResponse
+	// placeholder for future response values
 }
 
 // ActionGroupsClientUpdateResponse contains the response from method ActionGroupsClient.Update.